+++ src/arize/experimental/integrations/whylabs/__init__.py
@@ -1,5 +1,6 @@
+from arize.experimental.integrations.whylabs.client import IntegrationClient
 from arize.experimental.integrations.whylabs.generator import (
     WhylabsProfileAdapter,
 )
 
-__all__ = ["WhylabsProfileAdapter"]
+__all__ = ["WhylabsProfileAdapter", "IntegrationClient"]